refactor tauri structure
