+++ packages/synthchron-tauri/src/components/editor/LeftSidebar.tsx
