+++ packages/synthchron-tauri/src/components/editor/TabbedDrawer.tsx
