+++ packages/synthchron-tauri/src/components/editor/processModels/petriNet/customNodes/PlaceNode.tsx
