+++ packages/synthchron-tauri/src/components/editor/processModels/petriNet/petriNetFlowConfig.ts
