+++ packages/synthchron-tauri/src/utils/checkRoom.ts
