+++ packages/synthchron-tauri/src/utils/flowTransformer.ts
@@ -1,6 +1,6 @@
 import { ProcessModel } from '../../../synthchron-simulator/src/types/processModel'
 
-import { EditorState } from './react-flow/flowStore/flowStore'
+import { EditorState } from './react-flow/editorStore/flowStore'
 
 export const transformFlowToSimulator = (flow: EditorState): ProcessModel => {
   return flow.processModelFlowConfig.serialize(
